refactor(environment): share package summary surfaces
