+++ src/components/environment/index.ts
@@ -6,3 +6,13 @@ export {
   UvIcon,
   type EnvironmentIconProps,
 } from "./icons";
+export {
+  EnvironmentPackageSummaryPanel,
+  type EnvironmentPackageSummaryPanelProps,
+} from "./EnvironmentPackageSummaryPanel";
+export {
+  notebookMetadataToPackageViewModel,
+  type NotebookPackageManager,
+  type NotebookPackageSection,
+  type NotebookPackageViewModel,
+} from "./package-view-model";
